@@ -1,5 +1,8 @@
 // This is a script for deploying your contracts. You can adapt it to deploy
 // yours, or create new ones.
+
+const path = require("path");
+
 async function main() {
   // This is just a convenience check
   if (network.name === "hardhat") {
@@ -31,21 +34,21 @@ async function main() {
 
 function saveFrontendFiles(token) {
   const fs = require("fs");
-  const contractsDir = __dirname + "/../frontend/src/contracts";
+  const contractsDir = path.join(__dirname, "..", "frontend", "src", "contracts");
 
   if (!fs.existsSync(contractsDir)) {
     fs.mkdirSync(contractsDir);
   }
 
   fs.writeFileSync(
-    contractsDir + "/contract-address.json",
+    path.join(contractsDir, "contract-address.json"),
     JSON.stringify({ Token: token.address }, undefined, 2)
   );
 
   const TokenArtifact = artifacts.readArtifactSync("Token");
 
   fs.writeFileSync(
-    contractsDir + "/Token.json",
+    path.join(contractsDir, "Token.json"),
     JSON.stringify(TokenArtifact, null, 2)
   );
 }
