@@ -1,4 +1,4 @@
-// This is an exmaple test file. Hardhat will run every *.js file in `test/`,
+// This is an example test file. Hardhat will run every *.js file in `test/`,
 // so feel free to add new ones.
 
 // Hardhat tests are normally written with Mocha and Chai.
@@ -6,62 +6,59 @@
 // We import Chai to use its asserting functions here.
 const { expect } = require("chai");
 
-// `describe` is a Mocha function that allows you to organize your tests. It's
-// not actually needed, but having your tests organized makes debugging them
-// easier. All Mocha functions are available in the global scope.
+// We use `loadFixture` to share common setups (or fixtures) between tests.
+// Using this simplifies your tests and makes them run faster, by taking
+// advantage or Hardhat Network's snapshot functionality.
+const { loadFixture } = require("@nomicfoundation/hardhat-network-helpers");
 
-// `describe` recieves the name of a section of your test suite, and a callback.
-// The callback must define the tests of that section. This callback can't be
-// an async function.
+// `describe` is a Mocha function that allows you to organize your tests.
+// Having your tests organized makes debugging them easier. All Mocha
+// functions are available in the global scope.
+//
+// `describe` receives the name of a section of your test suite, and a
+// callback. The callback must define the tests of that section. This callback
+// can't be an async function.
 describe("Token contract", function () {
-  // Mocha has four functions that let you hook into the the test runner's
-  // lifecyle. These are: `before`, `beforeEach`, `after`, `afterEach`.
-
-  // They're very useful to setup the environment for tests, and to clean it
-  // up after they run.
-
-  // A common pattern is to declare some variables, and assign them in the
-  // `before` and `beforeEach` callbacks.
-
-  let Token;
-  let hardhatToken;
-  let owner;
-  let addr1;
-  let addr2;
-  let addrs;
-
-  // `beforeEach` will run before each test, re-deploying the contract every
-  // time. It receives a callback, which can be async.
-  beforeEach(async function () {
+  // We define a fixture to reuse the same setup in every test. We use
+  // loadFixture to run this setup once, snapshot that state, and reset Hardhat
+  // Network to that snapshot in every test.
+  async function deployTokenFixture() {
     // Get the ContractFactory and Signers here.
-    Token = await ethers.getContractFactory("Token");
-    [owner, addr1, addr2, ...addrs] = await ethers.getSigners();
+    const Token = await ethers.getContractFactory("Token");
+    const [owner, addr1, addr2] = await ethers.getSigners();
 
     // To deploy our contract, we just have to call Token.deploy() and await
     // for it to be deployed(), which happens onces its transaction has been
     // mined.
-    hardhatToken = await Token.deploy();
+    const hardhatToken = await Token.deploy();
 
-    // We can interact with the contract by calling `hardhatToken.method()`
     await hardhatToken.deployed();
-  });
+
+    // Fixtures can return anything you consider useful for your tests
+    return { Token, hardhatToken, owner, addr1, addr2 };
+  }
 
   // You can nest describe calls to create subsections.
   describe("Deployment", function () {
     // `it` is another Mocha function. This is the one you use to define your
     // tests. It receives the test name, and a callback function.
-
+//
     // If the callback function is async, Mocha will `await` it.
     it("Should set the right owner", async function () {
-      // Expect receives a value, and wraps it in an assertion objet. These
+      // We use loadFixture to setup our environment, and then assert that
+      // things went well
+      const { hardhatToken, owner } = await loadFixture(deployTokenFixture);
+
+      // Expect receives a value and wraps it in an assertion object. These
       // objects have a lot of utility methods to assert values.
 
-      // This test expects the owner variable stored in the contract to be equal
-      // to our Signer's owner.
+      // This test expects the owner variable stored in the contract to be
+      // equal to our Signer's owner.
       expect(await hardhatToken.owner()).to.equal(owner.address);
     });
 
     it("Should assign the total supply of tokens to the owner", async function () {
+      const { hardhatToken, owner } = await loadFixture(deployTokenFixture);
       const ownerBalance = await hardhatToken.balanceOf(owner.address);
       expect(await hardhatToken.totalSupply()).to.equal(ownerBalance);
     });
@@ -69,23 +66,32 @@ describe("Token contract", function () {
 
   describe("Transactions", function () {
     it("Should transfer tokens between accounts", async function () {
+      const { hardhatToken, owner, addr1, addr2 } = await loadFixture(deployTokenFixture);
       // Transfer 50 tokens from owner to addr1
-      await hardhatToken.transfer(addr1.address, 50);
-      const addr1Balance = await hardhatToken.balanceOf(
-        addr1.address
-      );
-      expect(addr1Balance).to.equal(50);
+      await expect(hardhatToken.transfer(addr1.address, 50))
+        .to.changeTokenBalances(hardhatToken, [owner, addr1], [-50, 50]);
 
       // Transfer 50 tokens from addr1 to addr2
       // We use .connect(signer) to send a transaction from another account
-      await hardhatToken.connect(addr1).transfer(addr2.address, 50);
-      const addr2Balance = await hardhatToken.balanceOf(
-        addr2.address
-      );
-      expect(addr2Balance).to.equal(50);
+      await expect(hardhatToken.connect(addr1).transfer(addr2.address, 50))
+        .to.changeTokenBalances(hardhatToken, [addr1, addr2], [-50, 50]);
     });
 
-    it("Should fail if sender doesn’t have enough tokens", async function () {
+    it("should emit Transfer events", async function () {
+      const { hardhatToken, owner, addr1, addr2 } = await loadFixture(deployTokenFixture);
+
+      // Transfer 50 tokens from owner to addr1
+      await expect(hardhatToken.transfer(addr1.address, 50))
+        .to.emit(hardhatToken, "Transfer").withArgs(owner.address, addr1.address, 50)
+
+      // Transfer 50 tokens from addr1 to addr2
+      // We use .connect(signer) to send a transaction from another account
+      await expect(hardhatToken.connect(addr1).transfer(addr2.address, 50))
+        .to.emit(hardhatToken, "Transfer").withArgs(addr1.address, addr2.address, 50)
+    });
+
+    it("Should fail if sender doesn't have enough tokens", async function () {
+      const { hardhatToken, owner, addr1 } = await loadFixture(deployTokenFixture);
       const initialOwnerBalance = await hardhatToken.balanceOf(
         owner.address
       );
@@ -100,34 +106,6 @@ describe("Token contract", function () {
       expect(await hardhatToken.balanceOf(owner.address)).to.equal(
         initialOwnerBalance
       );
-    });
-
-    it("Should update balances after transfers", async function () {
-      const initialOwnerBalance = await hardhatToken.balanceOf(
-        owner.address
-      );
-
-      // Transfer 100 tokens from owner to addr1.
-      await hardhatToken.transfer(addr1.address, 100);
-
-      // Transfer another 50 tokens from owner to addr2.
-      await hardhatToken.transfer(addr2.address, 50);
-
-      // Check balances.
-      const finalOwnerBalance = await hardhatToken.balanceOf(
-        owner.address
-      );
-      expect(finalOwnerBalance).to.equal(initialOwnerBalance - 150);
-
-      const addr1Balance = await hardhatToken.balanceOf(
-        addr1.address
-      );
-      expect(addr1Balance).to.equal(100);
-
-      const addr2Balance = await hardhatToken.balanceOf(
-        addr2.address
-      );
-      expect(addr2Balance).to.equal(50);
     });
   });
 });
